@@ -2,33 +2,10 @@
   "name": "client",
   "version": "0.1.0",
   "engines": {
-    "node": "10.15.0",
-    "npm": "6.5.0"
+    "node": "16.13.2",
+    "npm": "8.9.0"
   },
   "private": true,
-  "dependencies": {
-    "@material-ui/core": "^3.6.1",
-    "@material-ui/icons": "^3.0.1",
-    "@types/jwt-decode": "^2.2.1",
-    "@types/react-copy-to-clipboard": "^4.2.6",
-    "@types/react-redux": "^6.0.11",
-    "apexcharts": "^2.5.1",
-    "axios": "^0.18.0",
-    "jwt-decode": "^2.2.0",
-    "omit-empty": "^1.0.0",
-    "react": "^16.6.1",
-    "react-apexcharts": "^1.2.1",
-    "react-copy-to-clipboard": "^5.0.1",
-    "react-dom": "^16.6.1",
-    "react-redux": "^5.1.0",
-    "react-router-dom": "^4.3.1",
-    "react-scripts": "^2.1.3",
-    "react-star-rating-component": "^1.4.1",
-    "redux": "^4.0.1",
-    "redux-thunk": "^2.3.0",
-    "uuid-token-generator": "^1.0.0",
-    "node-sass": "^4.10.0"
-  },
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
@@ -38,11 +15,37 @@
   "eslintConfig": {
     "extends": "react-app"
   },
-  "proxy": "http://localhost:5000",
   "browserslist": [
     ">0.2%",
     "not dead",
     "not ie <= 11",
     "not op_mini all"
-  ]
+  ],
+  "dependencies": {
+    "@material-ui/core": "^4.12.4",
+    "@material-ui/icons": "^4.11.3",
+    "@types/jwt-decode": "^3.1.0",
+    "@types/react-copy-to-clipboard": "^5.0.4",
+    "@types/react-redux": "^7.1.25",
+    "apexcharts": "^3.37.1",
+    "axios": "^1.3.4",
+    "http-proxy-middleware": "^2.0.6",
+    "jwt-decode": "^3.1.2",
+    "node-sass": "^7.0.3",
+    "omit-empty": "^1.0.0",
+    "react": "^18.2.0",
+    "react-apexcharts": "^1.4.0",
+    "react-copy-to-clipboard": "^5.1.0",
+    "react-dom": "^17.0.2",
+    "react-redux": "^7.2.8",
+    "react-router-dom": "^6.8.2",
+    "react-scripts": "^5.0.1",
+    "react-star-rating-component": "^1.4.1",
+    "redux": "^4.2.1",
+    "redux-thunk": "^2.4.2",
+    "uuid": "^9.0.0"
+  },
+  "browser": {
+    "crypto": false
+  }
 }
