@@ -1,30 +1,30 @@
 {
-  "name": "medx",
-  "version": "1.0.0",
-  "description": "",
-  "main": "server.js",
-  "engines": {
-    "node": "10.15.0"
-  },
-  "scripts": {
-    "start" : "node server.js",
-    "server": "nodemon server.js",
-    "client": "npm start --prefix client",
-    "dev": "concurrently \"npm run server\" \"npm run client\"",
-    "heroku-postbuild": "cd client && npm install --only=dev && npm install && npm run build"
-  },
-  "author": "kgchicobava",
-  "license": "ISC",
-  "dependencies": {
-    "bcryptjs": "^2.4.3",
-    "body-parser": "^1.18.3",
-    "concurrently": "^4.0.1",
-    "express": "^4.16.4",
-    "jsonwebtoken": "^8.3.0",
-    "mongoose": "^5.3.10",
-    "omit-empty": "^1.0.0",
-    "passport": "^0.4.0",
-    "passport-jwt": "^4.0.0",
-    "validator": "^10.9.0"
+    "name": "medx",
+    "version": "1.0.0",
+    "description": "",
+    "main": "server.js",
+    "engines": {
+      "node": "16.13.2"
+    },
+    "scripts": {
+      "start" : "node server.js",
+      "server": "nodemon server.js",
+      "client": "npm start --prefix client",
+      "dev": "concurrently \"npm run server\" \"npm run client\"",
+      "heroku-postbuild": "cd client && npm install --only=dev && npm install && npm run build"
+    },
+    "author": "kgchicobava",
+    "license": "ISC",
+    "dependencies": {
+      "bcryptjs": "^2.4.3",
+      "body-parser": "^1.18.3",
+      "concurrently": "^4.0.1",
+      "express": "^4.16.4",
+      "jsonwebtoken": "^8.3.0",
+      "mongoose": "^5.3.10",
+      "omit-empty": "^1.0.0",
+      "passport": "^0.4.0",
+      "passport-jwt": "^4.0.0",
+      "validator": "^10.9.0"
+    }
   }
-}
