@@ -1,22 +1,17 @@
 import React from "react";
-import { Route, Redirect } from "react-router-dom";
+import { Route, Navigate } from "react-router-dom";
 import { connect } from "react-redux";
+import withRouter from "../../withRouter";
 import PropTypes from "prop-types";
 
-const PrivateRoute = ({ component: Component, auth, ...rest }) => (
-	<Route
-		{...rest}
-		render={props =>
-			auth.isDoctorAuthenticated === true ? (
-				<Component {...props} />
-			) : (
-				<Redirect to="/login" />
-			)
-		}
-	/>
-);
+function DoctorProtectedRoute({ auth, children }) {
+    if (!auth.isDoctorAuthenticated) {
+        return <Navigate to="/login" replace />;
+      }
+      return children;
+}
 
-PrivateRoute.propTypes = {
+DoctorProtectedRoute.propTypes = {
 	auth: PropTypes.object.isRequired
 };
 
@@ -24,4 +19,28 @@ const mapStateToProps = state => ({
 	auth: state.auth
 });
 
-export default connect(mapStateToProps)(PrivateRoute);
+let PrivateDoctorRoute = connect(mapStateToProps)(withRouter(DoctorProtectedRoute));
+export default PrivateDoctorRoute;
+
+// const PrivateRoute = ({ component: Component, auth, ...rest }) => (
+// 	<Route
+// 		{...rest}
+// 		component={props =>
+// 			auth.isDoctorAuthenticated === true ? (
+// 				<Component {...props} />
+// 			) : (
+// 				<Navigate to="/login" />
+// 			)
+// 		}
+// 	/>
+// );
+
+// PrivateRoute.propTypes = {
+// 	auth: PropTypes.object.isRequired
+// };
+
+// const mapStateToProps = state => ({
+// 	auth: state.auth
+// });
+
+// export default connect(mapStateToProps)(PrivateRoute);
