@@ -1,22 +1,17 @@
 import React from "react";
-import { Route, Redirect } from "react-router-dom";
+import { Route, Navigate } from "react-router-dom";
 import { connect } from "react-redux";
 import PropTypes from "prop-types";
+import withRouter from "../../withRouter";
 
-const PrivateRoute = ({ component: Component, auth, ...rest }) => (
-	<Route
-		{...rest}
-		render={props =>
-			auth.isPatientAuthenticated === true ? (
-				<Component {...props} />
-			) : (
-				<Redirect to="/login" />
-			)
-		}
-	/>
-);
+function PatientProtectedRoute({ auth, children }) {
+    if (!auth.isPatientAuthenticated) {
+        return <Navigate to="/login" replace />;
+      }
+      return children;
+}
 
-PrivateRoute.propTypes = {
+PatientProtectedRoute.propTypes = {
 	auth: PropTypes.object.isRequired
 };
 
@@ -24,4 +19,31 @@ const mapStateToProps = state => ({
 	auth: state.auth
 });
 
-export default connect(mapStateToProps)(PrivateRoute);
+let PrivatePatientRoute = connect(mapStateToProps)(withRouter(PatientProtectedRoute));
+export default PrivatePatientRoute ; 
+
+// const PrivateRoute = ({ component: Component, auth, ...rest }) => {
+// 	console.log(rest);
+// 	console.log("component: ", Component);
+// 	if(!auth.isPatientAuthenticated){
+// 					<Navigate to="/login" />
+// 	}
+// 	return (
+// 		<Route
+// 			{...rest}
+// 			render={props =>
+// 				<Component {...props} />
+// 			}
+// 		/>
+// 	)
+// 	};
+
+// PrivateRoute.propTypes = {
+// 	auth: PropTypes.object.isRequired
+// };
+
+// const mapStateToProps = state => ({
+// 	auth: state.auth
+// });
+
+// export default connect(mapStateToProps)(withRouter(PrivateRoute));
