@@ -3,20 +3,19 @@
 	@imported in App
 */
 import React, { Component } from "react";
+import withRouter from "../../withRouter";
 // Actions
 import ProfileActions from "../app-bar/ProfileActions";
 // Component
 import PatientTabs from "./PatientTabs";
 
-class PatientDashboard extends Component {
-	render() {
-		return (
-			<div>
-				<ProfileActions userRole="Patient" />
-				<PatientTabs />
-			</div>
-		);
-	}
+function PatientHomepage(){
+	return (
+		<div>
+			<ProfileActions userRole="Patient" />
+			<PatientTabs />
+		</div>
+	);
 }
 
-export default PatientDashboard;
+export default withRouter(PatientHomepage);
